Fix handshake delta LP64 divergence on 64-bit POSIX
The time-sync handshake used `long` for the delta arithmetic in
DESC::SendHandshake / HandshakeProcess. On LLP64 (Win64) long is 32-bit,
so `(long)(dwCurTime - dwTime)/2` reinterprets the wrapped unsigned diff
as signed and the `if (lNewDelta < 0)` recovery fires correctly. On LP64
(FreeBSD amd64) long is 64-bit: the same expression zero-extends the
32-bit unsigned difference to a huge positive value, the recovery never
runs, the delta diverges, and the handshake hits the 32-retry limit and
closes the connection.

Locally on Windows the sub-millisecond RTT lets round 1 converge before
the correction path is ever reached, hiding the bug. Through a higher-
latency link (the FreeBSD server reached via qemu user-networking), the
round-trip intermittently exceeds the 50ms bias window, the correction
path runs, and only there does LP64 diverge -- producing the flaky
connect / drop-on-select reported against the FreeBSD host.

Use `int` (32-bit on both data models) so the modular signed arithmetic
is identical everywhere, matching the long-working Win32 behaviour.

+++ Lead-Server-Source/game/src/desc.cpp
@@ -548,7 +548,7 @@ void DESC::StartHandshake(DWORD _handshake)
 	m_iHandshakeRetry = 0;
 }
 
-void DESC::SendHandshake(DWORD dwCurTime, long lNewDelta)
+void DESC::SendHandshake(DWORD dwCurTime, int lNewDelta)
 {
 	TPacketGCHandshake pack;
 
@@ -563,7 +563,7 @@ void DESC::SendHandshake(DWORD dwCurTime, long lNewDelta)
 	m_bHandshaking = true;
 }
 
-bool DESC::HandshakeProcess(DWORD dwTime, long lDelta, bool bInfiniteRetry)
+bool DESC::HandshakeProcess(DWORD dwTime, int lDelta, bool bInfiniteRetry)
 {
 	DWORD dwCurTime = get_dword_time();
 
@@ -593,7 +593,7 @@ bool DESC::HandshakeProcess(DWORD dwTime, long lDelta, bool bInfiniteRetry)
 		return true; 
 	}
 
-	long lNewDelta = (long) (dwCurTime - dwTime) / 2;
+	int lNewDelta = (int) (dwCurTime - dwTime) / 2;
 
 	if (lNewDelta < 0)
 	{
